+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
-All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+All notable changes to this project will be documented in this file.
 
 ### 0.0.40 (2024-03-28)
+1. 增加 utils.loadEnv() 加载系统环境变量